backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.3.5
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pytest==7.4.3
httpx==0.25.2

# Async support
aiofiles==23.2.1
websockets==12.0

# Audio processing
pydub==0.25.1
soundfile==0.12.1

# Storage and caching
boto3==1.33.6  # For S3 storage (optional)
redis==5.0.1   # For caching (optional)

# Monitoring and logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pytest-asyncio==0.21.1
pytest-cov==4.1.0
